ejercicio-1/requirements.txt
pytest
requests
sqlalchemy
pydantic
mysqlclient
python-multipart
python-jose[cryptography]
passlib[bcrypt]
Faker